mistakes correction

+++ sorting_algorithms.h
@@ -19,7 +19,7 @@ public:
     virtual void Sort(std::vector<int>& array) = 0;
 
     unsigned long long get_comparisons() const { return comparison; }
-    unsigned long get_permutations() const { return permutations; }
+    unsigned long long get_permutations() const { return permutations; }
 };
 
 class BubbleSort : public ISort {
